thread_test make pvl_func.py
-need button disable if you push.
